src/backend/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
black==22.3.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
croniter==1.3.7
debugpy==1.6.6
Deprecated==1.2.13
Django==4.0.3
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-ckeditor==6.5.1
django-cors-headers==3.12.0
django-editorjs-fields==0.2.7
django-enumchoicefield==3.0.1
django-fsm==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-parler==2.3
django-parler-rest==2.2
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-enumchoicefield==3.0.1
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
drf-yasg==1.20.0
flake8==5.0.4
filetype==1.2.0
idna==3.4
importlib-metadata==4.11.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonfield==2.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
Pillow==9.4.0
platformdirs==2.5.4
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
redis==4.3.4
requests==2.27.1
rq==1.11.1
qrcode==7.3.1
rq-scheduler==0.11.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zipp==3.8.0
django-rq==2.8.0
rq==1.13
